+++ src/Controllers/ControlPanel.js
@@ -1,6 +1,8 @@
 const Hyperparams = require("../Hyperparameters");
 const Modes = require("./ControlModes");
 const StatsPanel = require("../Stats/StatsPanel");
+const RandomOrganismGenerator = require("../Organism/RandomOrganismGenerator")
+const WorldConfig = require("../WorldConfig");
 
 class ControlPanel {
     constructor(engine) {
@@ -8,11 +10,10 @@ class ControlPanel {
         this.defineMinMaxControls();
         this.defineHotkeys();
         this.defineEngineSpeedControls();
-        this.defineGridSizeControls();
         this.defineTabNavigation();
         this.defineHyperparameterControls();
+        this.defineWorldControls();
         this.defineModeControls();
-        this.defineChallenges();
         this.fps = engine.fps;
         this.organism_record=0;
         this.env_controller = this.engine.env.controller;
@@ -114,38 +115,57 @@ class ControlPanel {
     defineEngineSpeedControls(){
         this.slider = document.getElementById("slider");
         this.slider.oninput = function() {
-            this.fps = this.slider.value
+            const max_fps = 300;
+            this.fps = this.slider.value;
+            if (this.fps>=max_fps) this.fps = 1000;
             if (this.engine.running) {
                 this.changeEngineSpeed(this.fps);
             }
-            $('#fps').text("Target FPS: "+this.fps);
+            let text = this.fps >= max_fps ? 'MAX' : this.fps;
+            $('#fps').text("Target FPS: "+text);
         }.bind(this);
+
         $('.pause-button').click(function() {
-            $('.pause-button').find("i").toggleClass("fa fa-pause");
-            $('.pause-button').find("i").toggleClass("fa fa-play");
-            this.paused = !this.paused;
-            if (this.engine.running) {
-                this.engine.stop();
-            }
-            else if (!this.engine.running){
-                this.engine.start(this.fps);
-            }
+            // toggle pause
+            this.setPaused(this.engine.running);
         }.bind(this));
+
         $('.headless').click(function() {
             $('.headless').find("i").toggleClass("fa fa-eye");
             $('.headless').find("i").toggleClass("fa fa-eye-slash");
-            if (Hyperparams.headless){
+            if (WorldConfig.headless){
                 $('#headless-notification').css('display', 'none');
                 this.engine.env.renderFull();
             }
             else {
                 $('#headless-notification').css('display', 'block');
             }
-            Hyperparams.headless = !Hyperparams.headless;
+            WorldConfig.headless = !WorldConfig.headless;
         }.bind(this));
     }
 
-    defineGridSizeControls() {
+    defineTabNavigation() {
+        this.tab_id = 'about';
+        var self = this;
+        $('.tabnav-item').click(function() {
+            $('.tab').css('display', 'none');
+            var tab = '#'+this.id+'.tab';
+            $(tab).css('display', 'grid');
+            $('.tabnav-item').removeClass('open-tab')
+            $('#'+this.id+'.tabnav-item').addClass('open-tab');
+            self.engine.organism_editor.is_active = (this.id == 'editor');
+            self.stats_panel.stopAutoRender();
+            if (this.id === 'stats') {
+                self.stats_panel.startAutoRender();
+            }
+            else if (this.id === 'editor') {
+                self.editor_controller.refreshDetailsPanel();
+            }
+            self.tab_id = this.id;
+        });
+    }
+
+    defineWorldControls() {
         $('#fill-window').change(function() {
             if (this.checked)
                 $('.col-row-input').css('display' ,'none');
@@ -168,24 +188,23 @@ class ControlPanel {
             this.stats_panel.reset();
             
         }.bind(this));
-    }
 
-    defineTabNavigation() {
-        this.tab_id = 'about';
-        var self = this;
-        $('.tabnav-item').click(function() {
-            $('.tab').css('display', 'none');
-            var tab = '#'+this.id+'.tab';
-            $(tab).css('display', 'grid');
-            self.engine.organism_editor.is_active = (this.id == 'editor');
-            self.stats_panel.stopAutoRender();
-            if (this.id === 'stats') {
-                self.stats_panel.startAutoRender();
-            }
-            else if (this.id === 'editor') {
-                self.editor_controller.refreshDetailsPanel();
-            }
-            self.tab_id = this.id;
+        $('#auto-reset').change(function() {
+            WorldConfig.auto_reset = this.checked;
+        });
+        $('#auto-pause').change(function() {
+            WorldConfig.auto_pause = this.checked;
+        });
+        $('#clear-walls-reset').change(function() {
+            WorldConfig.clear_walls_on_reset = this.checked;
+        });
+        $('#reset-with-editor-org').click( () => {
+            let env = this.engine.env;
+            if (!env.reset(true, false)) return;
+            let center = env.grid_map.getCenter();
+            let org = this.editor_controller.env.getCopyOfOrg();
+            this.env_controller.add_new_species = true;
+            this.env_controller.dropOrganism(org, center[0], center[1])
         });
     }
 
@@ -197,11 +216,8 @@ class ControlPanel {
             Hyperparams.lifespanMultiplier = $('#lifespan-multiplier').val();
         }.bind(this));
 
-        $('#mover-rot').change(function() {
-            Hyperparams.moversCanRotate = this.checked;
-        });
-        $('#offspring-rot').change(function() {
-            Hyperparams.offspringRotate = this.checked;
+        $('#rot-enabled').change(function() {
+            Hyperparams.rotationEnabled = this.checked;
         });
         $('#insta-kill').change(function() {
             Hyperparams.instaKill = this.checked;
@@ -211,6 +227,9 @@ class ControlPanel {
         });
         $('#food-drop-rate').change(function() {
             Hyperparams.foodDropProb = $('#food-drop-rate').val();
+        });
+        $('#extra-mover-cost').change(function() {
+            Hyperparams.extraMoverFoodCost = parseInt($('#extra-mover-cost').val());
         });
 
         $('#evolved-mutation').change( function() {
@@ -252,14 +271,40 @@ class ControlPanel {
         $('#reset-rules').click(() => {
             this.setHyperparamDefaults();
         });
+        $('#save-controls').click(() => {
+            let data = "data:text/json;charset=utf-8," + encodeURIComponent(JSON.stringify(Hyperparams));
+            let downloadEl = document.getElementById('download-el');
+            downloadEl.setAttribute("href", data);
+            downloadEl.setAttribute("download", "controls.json");
+            downloadEl.click();
+        });
+        $('#load-controls').click(() => {
+            $('#upload-el').click();
+        });
+        $('#upload-el').change((e)=>{
+            let files = e.target.files;
+            if (!files.length) {return;};
+            let reader = new FileReader();
+            reader.onload = (e) => {
+                let result=JSON.parse(e.target.result);
+                Hyperparams.loadJsonObj(result);
+                this.updateHyperparamUIValues();
+                // have to clear the value so change() will be triggered if the same file is uploaded again
+                $('#upload-el')[0].value = '';
+            };
+            reader.readAsText(files[0]);
+        });
     }
 
     setHyperparamDefaults() {
         Hyperparams.setDefaults();
+        this.updateHyperparamUIValues();
+    }
+
+    updateHyperparamUIValues(){
         $('#food-prod-prob').val(Hyperparams.foodProdProb);
         $('#lifespan-multiplier').val(Hyperparams.lifespanMultiplier);
-        $('#mover-rot').prop('checked', Hyperparams.moversCanRotate);
-        $('#offspring-rot').prop('checked', Hyperparams.offspringRotate);
+        $('#rot-enabled').prop('checked', Hyperparams.rotationEnabled);
         $('#insta-kill').prop('checked', Hyperparams.instaKill);
         $('#evolved-mutation').prop('checked', !Hyperparams.useGlobalMutability);
         $('#add-prob').val(Hyperparams.addProb);
@@ -268,6 +313,7 @@ class ControlPanel {
         $('#movers-produce').prop('checked', Hyperparams.moversCanProduce);
         $('#food-blocks').prop('checked', Hyperparams.foodBlocksReproduction);
         $('#food-drop-rate').val(Hyperparams.foodDropProb);
+        $('#extra-mover-cost').val(Hyperparams.extraMoverFoodCost);
         $('#look-range').val(Hyperparams.lookRange);
 
         if (!Hyperparams.useGlobalMutability) {
@@ -311,7 +357,6 @@ class ControlPanel {
             $('.edit-mode-btn').removeClass('selected');
             $('.'+this.id).addClass('selected');
         });
-
         $('.reset-view').click( function(){
             this.env_controller.resetView();
         }.bind(this));
@@ -324,19 +369,24 @@ class ControlPanel {
         $('#clear-env').click( () => {
             env.reset(true, false);
             this.stats_panel.reset();
-            env.auto_reset = false;
-            $('#auto-reset').prop('checked', false);;
         });
-        $('#auto-reset').change(function() {
-            env.auto_reset = this.checked;
-        });
+        $('#random-walls').click( function() {
+            this.env_controller.randomizeWalls();
+        }.bind(this));
         $('#clear-walls').click( function() {
             this.engine.env.clearWalls();
         }.bind(this));
         $('#clear-editor').click( function() {
             this.engine.organism_editor.clear();
             this.editor_controller.setEditorPanel();
-        }.bind(this))
+        }.bind(this));
+        $('#generate-random').click( function() {
+            this.engine.organism_editor.createRandom();
+            this.editor_controller.refreshDetailsPanel();
+        }.bind(this));
+        $('.reset-random').click( function() {
+            this.engine.organism_editor.resetWithRandomOrgs(this.engine.env);
+        }.bind(this));
 
         window.onbeforeunload = function (e) {
             e = e || window.event;
@@ -348,11 +398,22 @@ class ControlPanel {
         };
     }
 
-    defineChallenges() {
-        $('.challenge-btn').click(function() {
-            $('#challenge-title').text($(this).text());
-            $('#challenge-description').text($(this).val());
-        });
+    setPaused(paused) {
+
+        if (paused) {
+
+            $('.pause-button').find("i").removeClass("fa-pause");
+            $('.pause-button').find("i").addClass("fa-play");
+            if (this.engine.running) 
+                this.engine.stop();
+        }
+        else if (!paused) {
+            
+            $('.pause-button').find("i").addClass("fa-pause");
+            $('.pause-button').find("i").removeClass("fa-play");
+            if (!this.engine.running)
+                this.engine.start(this.fps);
+        }
     }
 
     setMode(mode) {
@@ -383,7 +444,7 @@ class ControlPanel {
     }
 
     updateHeadlessIcon(delta_time) {
-        if (this.paused)
+        if (!this.engine.running)
             return;
         const min_opacity = 0.4;
         var op = this.headless_opacity + (this.opacity_change_rate*delta_time/1000);
@@ -403,7 +464,7 @@ class ControlPanel {
         $('#fps-actual').text("Actual FPS: " + Math.floor(this.engine.actual_fps));
         $('#reset-count').text("Auto reset count: " + this.engine.env.reset_count);
         this.stats_panel.updateDetails();
-        if (Hyperparams.headless)
+        if (WorldConfig.headless)
             this.updateHeadlessIcon(delta_time);
 
     }