+++ src/Organism/Cell/GridCell.js
@@ -5,7 +5,7 @@ const Hyperparams = require("../../Hyperparameters");
 class Cell{
     constructor(state, col, row, x, y){
         this.owner = null; // owner organism
-        this.cell_owner = null; // owner cell of ^that organism
+        this.cell_owner = null; // specific body cell of the owner organism that occupies this grid cell
         this.setType(state);
         this.col = col;
         this.row = row;