+++ Changelog.md
@@ -1,5 +1,32 @@
 # Changelog
 
+## 1.0.2 (12/21/2021)
+
+### UI Enhancements:
+- New tab "World Controls"
+    - Relocated grid controls and auto reset to this tab
+    - Button to generate random walls with perlin noise
+    - Button to reset the environment with many randomly generated organisms
+    - Option to not clear walls on reset
+    - Option to pause on total extinction
+- "Simulation controls" tab renamed to "Evolution Controls"
+- Button to save/load Evolution Controls in a `.json` file
+- Button to randomize the organism in the editor window
+- Can now use drag view tool while rendering is off
+- Reorganized "About" tab and left panel, embedded explanation video
+
+### Simulation Enhancements:
+- New evolution control `Extra Mover Reproduction Cost`, which adds additional food cost for movers to reproduce
+- Combined `Movers can rotate` and `Offspring rotate` evolution controls into `Rotation enabled`
+- Fully max out simulation speed when slider is all the way to the right
+
+### Bug Fixes:
+- Armor is no longer ignored when checking for clear reproduction space
+- Chart data is now properly loaded/discarded when paused
+
+
+Thanks to contributors: @Chrispykins @M4YX0R
+
 ## 1.0.1 (12/4/2021)
 
 ### UI Enhancements: